Added all units and buildings. Basic balance.
Changed unit attack range to "double" for better gameplay.
Units are now constucted facing away from the building that produced them.
Launchers drop their goals if enemy comes into melee range.

+++ dune-server/src/main/java/lv/k2611a/util/AStar.java
@@ -22,7 +22,7 @@ public class AStar {
     public static int PATH_LENGTH = 20;
     public static double HARVESTER_HEURISTICS = 2.0;
 
-    public List<Node> calcPathEvenIfBlocked(Unit unit, Map map, int toX, int toY, int requiredRangeToGoal) {
+    public List<Node> calcPathEvenIfBlocked(Unit unit, Map map, int toX, int toY, double requiredRangeToGoal) {
 
         if ((unit.getX() == toX) && (unit.getY() == toY)) {
             return new ArrayList<Node>();
