+++ dune-server/src/test/java/lv/k2611a/domain/unitgoals/AttackMoveTest.java
@@ -122,13 +122,13 @@ public class AttackMoveTest {
         for (int coordY = 30; coordY <= 34; coordY+=2) {
             Building building = new Building();
             building.setType(BuildingType.FACTORY);
-            building.setX(killer.getX() - killer.getUnitType().getAttackRange());
+            building.setX(killer.getX() - (int)killer.getUnitType().getAttackRange());
             building.setY(coordY);
             building.setOwnerId(2);
             map.addBuilding(building);
             building = new Building();
             building.setType(BuildingType.FACTORY);
-            building.setX(killer.getX() - killer.getUnitType().getAttackRange() + 3);
+            building.setX(killer.getX() - (int)killer.getUnitType().getAttackRange() + 3);
             building.setY(coordY);
             building.setOwnerId(2);
             map.addBuilding(building);
@@ -165,14 +165,14 @@ public class AttackMoveTest {
         for (int coordY = 30; coordY <= 34; coordY+=2) {
             Building building = new Building();
             building.setType(BuildingType.FACTORY);
-            building.setX(killer.getX() - killer.getUnitType().getAttackRange());
+            building.setX(killer.getX() - (int)killer.getUnitType().getAttackRange());
             building.setY(coordY);
             building.setOwnerId(1);
             map.addBuilding(building);
 
             building = new Building();
             building.setType(BuildingType.FACTORY);
-            building.setX(killer.getX() - killer.getUnitType().getAttackRange() + 3);
+            building.setX(killer.getX() - (int)killer.getUnitType().getAttackRange() + 3);
             building.setY(coordY);
             building.setOwnerId(2);
             map.addBuilding(building);
